requirements.txt
python-dotenv==0.19.2
python-telegram-bot==13.8.1
google_cloud_dialogflow==2.10.0
